+++ f2837xd/cap/cap.h
@@ -19,7 +19,7 @@ SCOPED_ENUM_DECLARE_BEGIN(Peripheral) {
 } SCOPED_ENUM_DECLARE_END(Peripheral)
 
 
-const int peripheral_count = 6;
+const size_t peripheral_count = 6;
 
 
 namespace impl {
