+++ f2837xd/dac/dac.h
@@ -16,7 +16,7 @@ SCOPED_ENUM_DECLARE_BEGIN(Peripheral) {
 } SCOPED_ENUM_DECLARE_END(Peripheral)
 
 
-const int peripheral_count = 3;
+const size_t peripheral_count = 3;
 
 
 namespace impl {
