+++ f2837xd/i2c/i2c.h
@@ -16,7 +16,7 @@ SCOPED_ENUM_DECLARE_BEGIN(Peripheral) {
 } SCOPED_ENUM_DECLARE_END(Peripheral)
 
 
-const int peripheral_count = 2;
+const size_t peripheral_count = 2;
 
 
 SCOPED_ENUM_DECLARE_BEGIN(BitCount) {
